backend/requirements.txt
fastapi
openai
uvicorn[standard]
python-dotenv
pytest
